add missing documentation

+++ src/graphical_actions.rs
@@ -26,8 +26,10 @@ use serde::{Deserialize, Serialize};
 
 use crate::{anki_connect_send, Result};
 
+/// Parameters for the Card Browser dialog.
 #[derive(Default, Debug, Clone, PartialEq, Serialize)]
 pub struct GuiBrowseParams {
+    /// The search query for the Card Browser dialog.
     pub query: String,
 }
 
@@ -44,35 +46,48 @@ pub async fn gui_selected_notes() -> Result<Vec<usize>> {
     anki_connect_send::<(), _>("guiSelectedNotes", None).await
 }
 
+/// Parameters for adding cards using the Add Cards dialog.
 #[derive(Default, Debug, Clone, PartialEq, Serialize)]
 pub struct GuiAddCardsParams {
+    /// The note to add using the Add Cards dialog.
     pub note: GuiAddCardsNote,
 }
 
+/// A note for adding cards using the Add Cards dialog.
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct GuiAddCardsNote {
+    /// The deck name for the note.
     pub deck_name: String,
+    /// The model name for the note.
     pub model_name: String,
+    /// The fields of the note.
     pub fields: GuiAddCardsNoteFields,
+    /// The tags of the note.
     pub tags: Vec<String>,
+    /// The pictures attached to the note.
     pub picture: Vec<GuiAddCardsNotePicture>,
 }
 
+/// Fields of the note for adding cards using the Add Cards dialog.
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
 pub struct GuiAddCardsNoteFields {
+    /// The "Text" field of the note.
     #[serde(rename = "Text")]
     pub text: String,
+    /// The "Extra" field of the note.
     #[serde(rename = "Extra")]
     pub extra: String,
 }
 
+/// Picture attached to the note for adding cards using the Add Cards dialog.
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
 pub struct GuiAddCardsNotePicture {
+    /// The URL of the picture.
     pub url: String,
+    /// The filename of the picture.
     pub filename: String,
+    /// The fields associated with the picture.
     pub fields: Vec<String>,
 }
 
@@ -85,8 +100,10 @@ pub async fn gui_add_cards(params: GuiAddCardsParams) -> Result<usize> {
     anki_connect_send("guiAddCards", Some(params)).await
 }
 
+/// Parameters for editing a note using the Edit dialog.
 #[derive(Default, Debug, Clone, PartialEq, Serialize)]
 pub struct GuiEditNoteParams {
+    /// The ID of the note to edit.
     pub note: usize,
 }
 
@@ -100,34 +117,50 @@ pub async fn gui_edit_note(params: GuiEditNoteParams) -> Result<()> {
     anki_connect_send("guiEditNote", Some(params)).await
 }
 
-// TODO: Not implemented yet
 #[derive(Default, Debug, Clone, PartialEq, Deserialize)]
 #[serde(rename_all = "camelCase")]
+/// Represents the current card being reviewed.
 pub struct GuiCurrentCard {
+    /// The answer of the current card.
     pub answer: String,
+    /// The question of the current card.
     pub question: String,
+    /// The deck name of the current card.
     pub deck_name: String,
+    /// The model name of the current card.
     pub model_name: String,
+    /// The field order of the current card.
     pub field_order: isize,
+    /// The fields of the current card.
     pub fields: GuiCurrentCardFields,
+    /// The template of the current card.
     pub template: String,
+    /// The ID of the current card.
     pub card_id: isize,
+    /// The buttons associated with the current card.
     pub buttons: Vec<isize>,
+    /// The next reviews for the current card.
     pub next_reviews: Vec<String>,
 }
 
+/// Represents the fields of the current card being reviewed.
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct GuiCurrentCardFields {
+    /// The front face of the fields.
     #[serde(rename = "Front")]
+    /// The back face of the fields.
     pub front: GuiCurrentCardFieldsFace,
     #[serde(rename = "Back")]
     pub back: GuiCurrentCardFieldsFace,
 }
 
+/// Represents a face of the fields of the current card being reviewed.
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct GuiCurrentCardFieldsFace {
+    /// The value of the face.
     pub value: String,
-    pub order: i64,
+    /// The order of the face.
+    pub order: isize,
 }
 
 /// Returns information about the current card or `null` if not in review mode.
@@ -156,6 +189,7 @@ pub async fn gui_show_answer() -> Result<()> {
     anki_connect_send::<(), _>("guiShowAnswer", None).await
 }
 
+/// Parameters for answering the current card.
 #[derive(Default, Debug, Clone, PartialEq, Serialize)]
 pub struct GuiAnswerCardParams {
     pub cards: String,
@@ -167,8 +201,10 @@ pub async fn gui_answer_card(params: GuiAnswerCardParams) -> Result<bool> {
     anki_connect_send("guiAnswerCard", Some(params)).await
 }
 
+/// Parameters for opening the Deck Overview dialog.
 #[derive(Default, Debug, Clone, PartialEq, Serialize)]
 pub struct GuiDeckOverviewParams {
+    /// The name of the deck.
     pub name: String,
 }
 
@@ -185,8 +221,10 @@ pub async fn gui_deck_browser() -> Result<()> {
     anki_connect_send::<(), _>("guiDeckBrowser", None).await
 }
 
+/// Parameters for starting a deck review.
 #[derive(Default, Debug, Clone, PartialEq, Serialize)]
 pub struct GuiDeckReviewParams {
+    /// The name of the deck.
     pub name: String,
 }
